+++ packages/trrack-vis/src/components/ProvVis.tsx
@@ -4,28 +4,33 @@
 /* eslint-disable guard-for-in */
 /* eslint-disable no-unused-vars */
 import {
-  Provenance, isChildNode, NodeID, Nodes, ProvenanceNode, StateNode, DiffNode,
+  Provenance,
+  isChildNode,
+  NodeID,
+  Nodes,
+  ProvenanceNode,
+  StateNode,
+  DiffNode,
 } from '@visdesignlab/trrack';
 import {
   HierarchyNode,
   stratify,
   symbol,
   symbolWye,
-  symbolCross, symbolCircle, symbolTriangle, symbolSquare, symbolDiamond, symbolStar,
+  symbolCross,
+  symbolCircle,
+  symbolTriangle,
+  symbolSquare,
+  symbolDiamond,
+  symbolStar,
 } from 'd3';
 
 import React, {
   ReactChild, useEffect, useState, useCallback,
 } from 'react';
 import { NodeGroup } from 'react-move';
-import { Popup } from 'semantic-ui-react';
+import { Popup, Tab } from 'semantic-ui-react';
 
-import {
-  Tabs, Tab, AppBar, Box, Typography, Paper,
-} from '@material-ui/core';
-
-import ShareIcon from '@material-ui/icons/Share';
-import BookmarkIcon from '@material-ui/icons/Bookmark';
 import { style } from 'typestyle';
 
 import { BundleMap } from '../Utils/BundleMap';
@@ -34,7 +39,6 @@ import findBundleParent from '../Utils/findBundleParent';
 import translate from '../Utils/translate';
 
 import UndoRedoButton from './UndoRedoButton';
-import BookmarkToggle from './BookmarkToggle';
 import BookmarkListView from './BookmarkListView';
 import { treeLayout } from '../Utils/TreeLayout';
 import BackboneNode from './BackboneNode';
@@ -73,8 +77,8 @@ interface ProvVisProps<T, S extends string, A> {
   annotationContent?: (nodeId: StateNode<T, S, A>) => ReactChild;
   undoRedoButtons?: boolean;
   bookmarkToggle?: boolean;
-  bookmarkListView?:boolean;
-  editAnnotations?: boolean
+  bookmarkListView?: boolean;
+  editAnnotations?: boolean;
   prov?: Provenance<T, S, A>;
   ephemeralUndo?: boolean;
 }
@@ -91,7 +95,7 @@ function ProvVis<T, S extends string, A>({
   current,
   changeCurrent,
   width = 400,
-  height = 2000,
+  height = 800,
   iconOnly = false,
   gutter = 15,
   backboneGutter = 20,
@@ -121,10 +125,6 @@ function ProvVis<T, S extends string, A>({
   const [annotationOpen, setAnnotationOpen] = useState(-1);
   const [tabsValue, setValue] = useState(0);
 
-  if (popupContent) {
-    throw new Error('popups are not supported');
-  }
-
   let list: string[] = [];
   const eventTypes = new Set<string>();
   for (const j in nodeMap) {
@@ -134,12 +134,20 @@ function ProvVis<T, S extends string, A>({
         eventTypes.add(child.metadata.eventType);
       }
 
-      if (child.actionType === 'Ephemeral' && child.children.length === 1 && (nodeMap[child.parent].actionType !== 'Ephemeral' || nodeMap[child.parent].children.length > 1)) {
-        const group:string[] = [];
+      if (
+        child.actionType === 'Ephemeral'
+        && child.children.length === 1
+        && (nodeMap[child.parent].actionType !== 'Ephemeral'
+          || nodeMap[child.parent].children.length > 1)
+      ) {
+        const group: string[] = [];
         let curr = child;
         while (curr.actionType === 'Ephemeral') {
           group.push(curr.id);
-          if (curr.children.length === 1 && nodeMap[curr.children[0]].actionType === 'Ephemeral') {
+          if (
+            curr.children.length === 1
+            && nodeMap[curr.children[0]].actionType === 'Ephemeral'
+          ) {
             curr = nodeMap[curr.children[0]] as DiffNode<T, S, A>;
           } else {
             break;
@@ -159,9 +167,13 @@ function ProvVis<T, S extends string, A>({
     list = list.concat(Object.keys(bundleMap));
   }
 
-  function setDefaultConfig<E extends string>(types:Set<string>): EventConfig<E> {
+  function setDefaultConfig<E extends string>(
+    types: Set<string>,
+  ): EventConfig<E> {
     const symbols = [
-      symbol().type(symbolStar).size(50),
+      symbol()
+        .type(symbolStar)
+        .size(50),
       symbol().type(symbolDiamond),
       symbol().type(symbolTriangle),
       symbol().type(symbolCircle),
@@ -214,7 +226,9 @@ function ProvVis<T, S extends string, A>({
     return conf;
   }
 
-  const [expandedClusterList, setExpandedClusterList] = useState<string[]>(Object.keys(bundleMap));
+  const [expandedClusterList, setExpandedClusterList] = useState<string[]>(
+    Object.keys(bundleMap),
+  );
 
   if (!eventConfig && eventTypes.size > 0 && eventTypes.size < 8) {
     eventConfig = setDefaultConfig<S>(eventTypes);
@@ -224,9 +238,7 @@ function ProvVis<T, S extends string, A>({
     setFirst(false);
   }, []);
 
-  const nodeList = Object.values(nodeMap).filter(
-    (d) => true,
-  );
+  const nodeList = Object.values(nodeMap).filter(() => true);
 
   const copyList = Array.from(nodeList);
 
@@ -332,7 +344,9 @@ function ProvVis<T, S extends string, A>({
   const stratifiedList: StratifiedList<T, S, A> = stratifiedTree.descendants();
   const stratifiedMap: StratifiedMap<T, S, A> = {};
 
-  stratifiedList.forEach((c) => { stratifiedMap[c.id!] = c; });
+  stratifiedList.forEach((c) => {
+    stratifiedMap[c.id!] = c;
+  });
   treeLayout(stratifiedMap, current, root);
 
   let maxHeight = 0;
@@ -402,22 +416,6 @@ function ProvVis<T, S extends string, A>({
   }
 
   const svgWidth = width;
-
-  // if (document.getElementById("globalG") !== null) {
-  //   if (
-  //     document
-  //       .getElementById("globalG")!
-  //       .getBoundingClientRect()
-  //       .width.valueOf() > svgWidth
-  //   ) {
-  //     console.log("in here");
-  //     svgWidth =
-  //       document
-  //         .getElementById("globalG")!
-  //         .getBoundingClientRect()
-  //         .width.valueOf() + 10;
-  //   }
-  // }
 
   const overflowStyle = {
     overflowX: 'auto',
@@ -674,76 +672,25 @@ function ProvVis<T, S extends string, A>({
     </div>
   );
 
-  const handleChange = useCallback((event: any, newValue: number) => {
-    setValue(newValue);
-  }, []);
+  const panes = [
+    {
+      menuItem: { key: 'Graph', icon: 'share alternate', content: 'Graph' },
+      render: () => <Tab.Pane attached={false}>{graphTabView}</Tab.Pane>,
+    },
+    {
+      menuItem: { key: 'Bookmarks/Annotations', icon: 'bookmark', content: 'Bookmarks/Annotations' },
+      render: () => <Tab.Pane attached={false}>{bookmarkTabView}</Tab.Pane>,
+    },
+  ];
 
   return (
     <div style={overflowStyle} className={container} id="prov-vis">
-      <Paper square style={tabsStyle}>
-        <Tabs
-          value={tabsValue}
-          onChange={handleChange}
-          indicatorColor="primary"
-          textColor="primary"
-          variant="fullWidth"
-          aria-label="full width tabs example"
-          style={tabsStyle}
-        >
-          <Tab
-            style={{ maxWidth: 135, minWidth: 135 }}
-            icon={<ShareIcon />}
-            {...a11yProps(0)}
-          />
-          <Tab
-            style={{ maxWidth: 135, minWidth: 135 }}
-            icon={<BookmarkIcon />}
-            {...a11yProps(1)}
-          />
-        </Tabs>
-      </Paper>
-      <TabPanel value={tabsValue} index={0}>
-        {graphTabView}
-      </TabPanel>
-      <TabPanel value={tabsValue} index={1}>
-        {bookmarkTabView}
-      </TabPanel>
+      <Tab menu={{ secondary: true, pointing: true }} panes={panes} />
     </div>
   );
 }
 
 export default ProvVis;
-
-function TabPanel(props:any) {
-  const {
-    children, value, index, ...other
-  } = props;
-
-  console.log(children);
-  console.log(value);
-  console.log(index);
-
-  return (
-    <div
-      role="tabpanel"
-      hidden={value !== index}
-      id={`simple-tabpanel-${index}`}
-      aria-labelledby={`simple-tab-${index}`}
-      {...other}
-    >
-      {value === index && (
-        <div>{children}</div>
-      )}
-    </div>
-  );
-}
-
-function a11yProps(index: number) {
-  return {
-    id: `full-width-tab-${index}`,
-    'aria-controls': `full-width-tabpanel-${index}`,
-  };
-}
 
 const container = style({
   alignItems: 'center',